+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Soundary"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,13 +22,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".FriendSearchActivity"
+            android:name=".search.FriendSearchActivity"
             android:exported="false" />
         <activity
-            android:name=".ShareFriendActivity"
+            android:name=".share.ShareFriendActivity"
             android:exported="false" />
         <activity
-            android:name=".ShareMusicActivity"
+            android:name=".share.ShareMusicActivity"
             android:exported="false" />
     </application>
