--- app/src/main/java/com/team22/soundary/FriendSearchActivity.kt
@@ -1,14 +1,0 @@
-package com.team22.soundary
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-
-class FriendSearchActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_friend_search)
-    }
-}
