--- app/src/main/java/com/team22/soundary/ProfileActivity.kt
@@ -1,4 +1,0 @@
-package com.team22.soundary
-
-class ProfileActivity {
-}
