+++ app/src/main/java/com/team22/soundary/profile/ProfileActivity.kt
@@ -1,0 +1,4 @@
+package com.team22.soundary.profile
+
+class ProfileActivity {
+}
