+++ app/src/main/java/com/team22/soundary/search/FriendSearchActivity.kt
@@ -1,11 +1,12 @@
-package com.team22.soundary;
+package com.team22.soundary.search
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.team22.soundary.R
 
-class ShareFriendActivity : AppCompatActivity() {
+class FriendSearchActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_share_friend)
+        setContentView(R.layout.activity_friend_search)
     }
 }
