+++ app/src/main/java/com/team22/soundary/share/ShareFriendActivity.kt
@@ -1,11 +1,12 @@
-package com.team22.soundary;
+package com.team22.soundary.share;
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.team22.soundary.R
 
-class ShareMusicActivity : AppCompatActivity() {
+class ShareFriendActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_share_music)
+        setContentView(R.layout.activity_share_friend)
     }
 }
