+++ app/src/main/java/com/team22/soundary/share/ShareMusicActivity.kt
@@ -1,0 +1,12 @@
+package com.team22.soundary.share;
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.team22.soundary.R
+
+class ShareMusicActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_share_music)
+    }
+}
