+++ app/src/main/res/layout/activity_share_music.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/share_padding"
-    tools:context=".ShareMusicActivity">
+    tools:context=".share.ShareMusicActivity">
 
     <TextView
         android:id="@+id/share_music_top_textview"
